+++ src/Driver.h
@@ -46,7 +46,7 @@ public:
   void initializeSolver();
   void solve();
   bool isSatisfiable();
-  std::string count(std::string var_name, int bound, bool count_less_than_or_equal_to_bound = true);
+  std::string count(std::string var_name, double bound, bool count_less_than_or_equal_to_bound = true);
   std::string count(int bound, bool count_less_than_or_equal_to_bound = true);
   void printResult(Solver::Value_ptr value, std::ostream& out);
   void inspectResult(Solver::Value_ptr value, std::string file_name);
